Main activity layout changed to allow FAB to exist above the fragments and an animation exists for it
