+++ app/src/main/java/acc/aviato/MainActivity.java
@@ -1,6 +1,7 @@
 package acc.aviato;
 
 
+import android.app.AlertDialog;
 import android.content.Intent;
 import android.location.Location;
 import android.os.Bundle;
@@ -10,16 +11,26 @@ import android.support.v4.view.ViewPager;
 import android.support.v7.app.AppCompatActivity;
 import android.view.Menu;
 import android.view.MenuItem;
+import android.view.View;
+import android.view.animation.AccelerateInterpolator;
+import android.view.animation.Animation;
+import android.view.animation.DecelerateInterpolator;
+import android.view.animation.ScaleAnimation;
+import android.view.animation.TranslateAnimation;
 import android.widget.TextView;
 
 import com.google.android.gms.common.api.GoogleApiClient;
 import com.google.android.gms.location.LocationServices;
 import com.google.android.gms.common.ConnectionResult;
+import com.melnykov.fab.FloatingActionButton;
+import com.parse.ParseUser;
 
 public class MainActivity extends AppCompatActivity {
 
     protected SectionsPageAdapter mSectionsPagerAdapter;
     protected ViewPager mViewPager;
+
+    FloatingActionButton mFab;
 
     @Override
     protected void onCreate(Bundle savedInstanceState) {
@@ -32,8 +43,42 @@ public class MainActivity extends AppCompatActivity {
         mViewPager.setAdapter(mSectionsPagerAdapter);
         mViewPager.setCurrentItem(0, false);    // Sets the default tab
 
+        mFab = (FloatingActionButton) findViewById(R.id.fab_feed);
+        mFab.setOnClickListener(new View.OnClickListener() {
+            @Override
+            public void onClick(View v) {
+                if (ParseUser.getCurrentUser() == null) {
+                    AlertDialog.Builder builder = new AlertDialog.Builder(MainActivity.this);
+                    builder.setMessage("You must be signed in to create an event.")
+                            .setTitle("Not signed in")
+                            .setPositiveButton(android.R.string.ok, null);
+                    AlertDialog alert = builder.create();
+                    alert.show();
+                } else {
+                    Intent intent = new Intent(MainActivity.this, CreateEventActivity.class);
+                    startActivity(intent);
+                }
+            }
+        });
+
         TabLayout tabLayout = (TabLayout) findViewById(R.id.main_tabs);
         tabLayout.setupWithViewPager(mViewPager);
+
+        tabLayout.setOnTabSelectedListener(new TabLayout.OnTabSelectedListener() {
+            @Override
+            public void onTabSelected(TabLayout.Tab tab) {
+                mViewPager.setCurrentItem(tab.getPosition());
+                animateFab(tab.getPosition());
+            }
+
+            @Override
+            public void onTabUnselected(TabLayout.Tab tab) {
+            }
+
+            @Override
+            public void onTabReselected(TabLayout.Tab tab) {
+            }
+        });
     }
 
 
@@ -65,47 +110,15 @@ public class MainActivity extends AppCompatActivity {
         return super.onOptionsItemSelected(item);
     }
 
-    /*public void createFilterDialog(){
-        AlertDialog.Builder builder = new AlertDialog.Builder(this);
-        // Add the buttons
-        builder.setPositiveButton(R.string.filter, new DialogInterface.OnClickListener() {
-            public void onClick(DialogInterface dialog, int id) {
-                // User clicked OK button
-            }
-        });
-        builder.setNegativeButton(R.string.cancel, new DialogInterface.OnClickListener() {
-            public void onClick(DialogInterface dialog, int id) {
-                // User cancelled the dialog
-            }
-        });
-        // Set other dialog properties
-        builder.setTitle("Filter Events");
-        builder.setView(R.layout.content_filter_dialog);
 
-        final AlertDialog dialog = builder.create();
+    protected void animateFab(final int position) {
+        // Creating a custom animation will allow more controlled animations
+        if (position != 0) {
+            mFab.hide();
+        } else {
+            mFab.show();
+        }
+    }
 
-
-        ParseQuery<ParseObject> comp = new ParseQuery<ParseObject>(ParseConstants.CLASS_TAGS);
-        comp.findInBackground(new FindCallback<ParseObject>() {
-            @Override
-            public void done(List<ParseObject> list, ParseException e) {
-                if (e == null) {
-                    //If I weren't lazy I would make a sorting algorithm here to sort this list by tag popularity, but I'm not even sure that would make a difference so I won't do it for now
-                    String[] tagFilters = new String[list.size()];
-                    for (int i = 0; i < list.size(); i++) {
-                        tagFilters[i]=list.get(i).get(ParseConstants.KEY_TAG_NAME).toString();
-                    }
-                    ArrayAdapter<String> adapter = new ArrayAdapter<String>(getBaseContext(),android.R.layout.simple_dropdown_item_1line,tagFilters);
-                    AutoCompleteTextView textView = (AutoCompleteTextView)dialog.findViewById(R.id.filter_autocomplete);
-                    textView.setAdapter(adapter);
-
-                }
-            }
-        });
-
-        // Create the AlertDialog
-
-        dialog.show();
-    }*/
 
 }
