+++ app/src/main/res/layout/fragment_favorites_activity.xml
@@ -13,7 +13,8 @@
         android:layout_alignParentRight="true"
         android:layout_alignParentStart="true"
         android:layout_alignParentTop="true"
-        android:padding="10dp" />
+        android:paddingLeft="10dp"
+        android:paddingRight="10dp" />
 
     <TextView
         android:id="@android:id/empty"
